All of the stuff below refers to the full solver pipeline--not just the milp solver.
Good news: the SparseAdjacencyMILP formulation gets solved way faster
than the dense AdjacencyMILP formulation.

Bad news: Problems 75 and 80 (and perhaps others) are resulting in a cyclic
project graph when I use the SparseAdjacencyMILP formulation. Strangely,
the same thing happens on problem 74 (but not 75 or 80) for the dense
AdjacencyMILP formulation. I have yet to figure out why...

Update: the file TaskGraphs/log.txt contains information about all
problem instances on which AdjacencyMILP or SparseAdjacencyMILP leads to
an error within the overall solution process. So far as I have seen,
these errors originate from cyclic project graph or optimization
timeout. The latter is irrelevant for debugging--it's just getting
caught by the same try-catch block that catches the other errors.

I previously noted that costs don't match between AdjacencyMILP and
SparseAdjacencyMILP on a few problem instances, but it appears that this
only happens when a cyclic project graph gets in the way. So it doesn't
seem like something to worry about after all.

Finally, the AssigmentMILP formulation can't get all the way through
problem1 in the test set. I need to identify what's causing the problem.