Great. Now none of the saved problem instances seem to be working. I shall investigate... The problem can be traced to the A_star heuristics. Since I cleaned out the starts and goals, all of the heuristics are dead. Fixing now. Problem solved! Just made sure to completely populate the perfect heuristics (this time I went all out and used the full distance matrix.)
I also discovered that the cyclic project graph errors were occurring
after many iterations of the MILP solver. I added a check to skip the
assignment phase if the project_schedule is invalid, and to return the
best_solution if the MILP solver iteration limit is exceeded. On
problems 74 and 75, which were failing for AdjacencyMILP and
SparseAdjacencyMILP, respectively, I found that the final solution has
an optimality gap of 1.0. Somehow an impossible-to-beat error has crept
in there.