For some reason, the algorithm still has an enormous optimality gap for replanning. I am not sure why. It currently works as follows:
- A project schedule and route plan are obtained from the solution of
  the first problem.
- The schedule is pruned below a certain "commit threshold", and the
  schedule and route plan are both effectively frozen at that threshold.
All schedule nodes that terminate prior to the commit threshold are
designated as "fixed", meaning that they do not need planning.
- All "assignments" are deleted from the pruned project schedule - NOTE:
  I HAVE NOT ACTUALLY DONE THIS YET...
- A new project schedule is added to the current schedule
- The resulting partially-assigned project schedule, along with the
  frozen solution and the planning cache, are passed into a base search
env.
- The base search env is passed into the solver.

A few things that I have specifically fixed:
- After the MILP is solved, update_project_schedule! does not modify the
  "fixed" vertices. It was previously doning so, which is not allowed.

One possibility: perhaps the algorithm is assuming that the RobotAT
nodes--which have been updated in the new schedule, start at t=0 instead
of t = cutoff time.
Actually, I think that is already taken care of...