Somehow the route plan is still thinking that the robot is in a different place at the beginning. Investigating now.
Looking at the schedule, it appears that the initial state of the robot
is being accurately set on each replan! iteration, but the ROBOT_AT node
is still appearing right at the beginning of the schedule (instead of
being inserted in the middle of the active GO, CARRY, COLLECT or DEPOSIT
node.

Hypothesis: The problem is that PIBT doesn't "skip" the nodes for which
plan_path=false, thus forcing the robot to start over from the
beginning.
- The hypothesis was at least partially correct. It has now been fixed
  so that update_planning_cache! and pibt_update_envs! are called in a
preprocessing step before PIBT starts marching. However, We are still
seeing funny changes in the route plan after replan! has been called.
The initial ROBOT_AT node always seems to stay in the schedule (as
opposed to being pruned away), and it's terminal time keeps increasing.
Perhaps the problem is being caused by update_env! or
update_planning_cache! passing a long path when updating a particular
node, which causes cache.tF for that node to be equal to the path
length.