Just realized that removing vertices from GridFactoryWorld also needs to involve an update to the vtx_map and vtx_list. How to do that?
FALSE ALARM! No need to worry about issues with removing vertices,
because we will only be removing edges. On that note, this might be a
good time to think about how to represent custom graphs depending on the
robot/team type (e.g., no-go zones are different for a DeliveryBot than
for a CleanUpBot).