Trying to decide if/how to refactor ProblemSpec. It has a ton of fields, most of which are only used in one place (the AssigmnentMILP solver).
Possible changes:
- Replace the current ProblemSpec with a parametric type that encodes
  EntityID=>(entity start time, entity initial state, etc.), for all
object and robots. Similar to what is already implemented with PathSpec.

I think I'll start by allowing the formulate_milp(AssignmentMILP,...)
method to access the base OperatingSchedule, as it does for the other
assignment models. This will make it possible to used the initial time
information from the schedule.